refactor(Combat): merge combat/combat-map-entity
 - this is yet another unnecessary level of indirection that made it hard to follow how the combat component hierarchy was set up.
 - move the code in combat-map.entity and combat-camera-behavior into the combat component itself.
