--- src/app/routes/combat/behaviors/combat-camera.behavior.ts
@@ -1,49 +1,0 @@
-/*
- Copyright (C) 2013-2020 by Justin DuJardin and Contributors
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- */
-import { Component } from '@angular/core';
-import { Rect } from '../../../../app/core/rect';
-import { CameraBehavior } from '../../../behaviors/camera-behavior';
-import { SceneView } from '../../../scene/scene-view';
-import { TileMap } from '../../../scene/tile-map';
-
-@Component({
-  selector: 'combat-camera-behavior',
-  template: `<ng-content></ng-content>`,
-})
-export class CombatCameraBehaviorComponent extends CameraBehavior {
-  host: TileMap;
-
-  connectBehavior(): boolean {
-    return super.connectBehavior() && this.host instanceof TileMap;
-  }
-
-  process(view: SceneView) {
-    if (!this.host || !view.context) {
-      super.process(view);
-      return;
-    }
-    const w: number = view.context.canvas.width;
-    const screenRect = new Rect(
-      0,
-      0,
-      view.context.canvas.width,
-      view.context.canvas.height
-    );
-    view.cameraScale = w > 1024 ? 6 : w > 768 ? 4 : w > 480 ? 3 : 2;
-    view.camera = view.screenToWorld(screenRect, view.cameraScale);
-    view.camera.point.x = this.host.bounds.extent.x / 2 - view.camera.extent.x / 2;
-  }
-}
