+++ src/app/routes/combat/behaviors/index.ts
@@ -1,3 +1,1 @@
 export * from './combat-action.behavior';
-export * from './combat-camera.behavior';
-export * from './combat-player-render.behavior';
