+++ src/app/routes/combat/combat-enemy.component.html
@@ -1,0 +1,5 @@
+<sprite-render-behavior
+  [icon]="(model$ | async)?.icon || null"
+></sprite-render-behavior>
+<combat-attack-behavior [combat]="combat"></combat-attack-behavior>
+<ng-content></ng-content>
