+++ src/app/routes/combat/combat-enemy.component.ts
@@ -9,13 +9,7 @@ import { CombatAttackBehaviorComponent } from './behaviors/actions/combat-attack
 
 @Component({
   selector: 'combat-enemy',
-  template: `
-    <sprite-render-behavior
-      [icon]="(model$ | async)?.icon || null"
-    ></sprite-render-behavior>
-    <combat-attack-behavior [combat]="combat"></combat-attack-behavior>
-    <ng-content></ng-content>
-  `,
+  templateUrl: 'combat-enemy.component.html',
 })
 export class CombatEnemyComponent
   extends GameEntityObject
