--- src/app/routes/combat/combat-map.entity.html
@@ -1,18 +1,0 @@
-<combat-camera-behavior></combat-camera-behavior>
-<combat-player
-  *ngFor="let p of (party$ | async); trackBy: combatTrackBy"
-  [icon]="p.icon"
-  [model]="p"
-  [scene]="scene"
-  [combat]="combat"
-  #partyMember
-></combat-player>
-<combat-enemy
-  *ngFor="let e of (enemies$ | async); trackBy: combatTrackBy"
-  [icon]="e.icon"
-  [model]="e"
-  [scene]="scene"
-  [combat]="combat"
-  #enemy
-></combat-enemy>
-<ng-content></ng-content>
