--- src/app/routes/combat/combat-map.entity.ts
@@ -1,196 +1,0 @@
-import {
-  AfterViewInit,
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  OnDestroy,
-  QueryList,
-  ViewChild,
-  ViewChildren,
-} from '@angular/core';
-import { Store } from '@ngrx/store';
-import { Observable, Subscription } from 'rxjs';
-import { distinctUntilChanged, map } from 'rxjs/operators';
-import { IPoint, Point } from '../../../app/core/point';
-import { ResourceManager } from '../../../app/core/resource-manager';
-import { TiledTMXResource } from '../../../app/core/resources/tiled/tiled-tmx.resource';
-import { AppState } from '../../app.model';
-import { SpriteComponent } from '../../behaviors/sprite.behavior';
-import { LoadingService } from '../../components/loading/loading.service';
-import { ITiledObject } from '../../core/resources/tiled/tiled.model';
-import { CombatantTypes, IEnemy, IPartyMember } from '../../models/base-entity';
-import { CombatState } from '../../models/combat/combat.model';
-import { CombatService } from '../../models/combat/combat.service';
-import {
-  getCombatEncounterEnemies,
-  getCombatEncounterParty,
-  sliceCombatState,
-} from '../../models/selectors';
-import { TileMapRenderer } from '../../scene/render/tile-map-renderer';
-import {
-  TileObjectRenderer,
-  TileRenderable,
-} from '../../scene/render/tile-object-renderer';
-import { Scene } from '../../scene/scene';
-import { SceneView } from '../../scene/scene-view';
-import { ISceneViewRenderer } from '../../scene/scene.model';
-import { TileMap } from '../../scene/tile-map';
-import { CombatCameraBehaviorComponent } from './behaviors/combat-camera.behavior';
-import { CombatEnemyComponent } from './combat-enemy.entity';
-import { CombatPlayerComponent } from './combat-player.entity';
-
-@Component({
-  selector: 'combat-map',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: './combat-map.entity.html',
-})
-export class CombatMapComponent
-  extends TileMap
-  implements AfterViewInit, OnDestroy, ISceneViewRenderer
-{
-  private _tileRenderer: TileMapRenderer = new TileMapRenderer();
-
-  @Input() scene: Scene;
-  @Input() combat: any; // _ TODO: This creates a nasty circular import situation
-
-  @ViewChild(CombatCameraBehaviorComponent)
-  camera: CombatCameraBehaviorComponent;
-
-  enemies$: Observable<IEnemy[]> = this.store
-    .select(getCombatEncounterEnemies)
-    .pipe(map((f) => f.toJS()));
-
-  party$: Observable<IPartyMember[]> = this.store
-    .select(getCombatEncounterParty)
-    .pipe(map((f) => f.toJS()));
-
-  encounter$: Observable<CombatState> = this.store.select(sliceCombatState);
-
-  @ViewChildren('partyMember') party: QueryList<CombatPlayerComponent>;
-  @ViewChildren('enemy') enemies: QueryList<CombatEnemyComponent>;
-
-  /** The {@see TiledTMXResource} map at the input url */
-  resource$: Observable<TiledTMXResource> = this.combatService.combatMap$.pipe(
-    distinctUntilChanged(),
-    map((result: TiledTMXResource) => {
-      this.setMap(result);
-      return result;
-    })
-  );
-
-  /** Features can be derived after a new map resource has been loaded */
-  features$: Observable<ITiledObject[]> = this.resource$.pipe(
-    map(() => {
-      return this.features?.objects || [];
-    })
-  );
-
-  constructor(
-    public combatService: CombatService,
-    public store: Store<AppState>,
-    public loadingService: LoadingService,
-    public loader: ResourceManager
-  ) {
-    super();
-  }
-
-  /**
-   * @internal used to make `resource$` observable hot. Can be removed if the template references an
-   * async binding for this at any point.
-   */
-  private _resourceSubscription: Subscription | null;
-
-  ngAfterViewInit(): void {
-    this.scene?.addObject(this);
-    this.addBehavior(this.camera);
-    // Whenever the underlying map resource changes, update party/enemy state from latest values.
-    this._resourceSubscription = this.resource$
-      .pipe(
-        map(() => {
-          this.party.forEach((p: CombatPlayerComponent, index: number) => {
-            p.setSprite(p.model.icon);
-            const battleSpawn = this.getFeature('p' + (index + 1)) as IPoint;
-            p.setPoint(new Point(battleSpawn.x / 16, battleSpawn.y / 16));
-          });
-          this.enemies.forEach((e: CombatEnemyComponent, index: number) => {
-            const battleSpawn = this.getFeature('e' + (index + 1)) as IPoint;
-            if (battleSpawn) {
-              e.setPoint(new Point(battleSpawn.x / 16, battleSpawn.y / 16));
-            }
-          });
-        })
-      )
-      .subscribe();
-  }
-
-  ngOnDestroy(): void {
-    this._resourceSubscription?.unsubscribe();
-    this.scene?.removeObject(this);
-    this.removeBehavior(this.camera);
-    this.destroy();
-  }
-
-  //
-  // ISceneViewRenderer
-  //
-  objectRenderer: TileObjectRenderer = new TileObjectRenderer();
-
-  beforeFrame(view: SceneView, elapsed: number) {
-    // Nope
-  }
-
-  /**
-   * Render all of the map feature components
-   */
-  renderFrame(view: SceneView, elapsed: number) {
-    this._tileRenderer.render(this, view);
-    this.party.forEach((component: CombatPlayerComponent) => {
-      this.objectRenderer.render(
-        component,
-        component.renderPoint || component.point,
-        view,
-        component.meta
-      );
-      const sprites = component.findBehaviors(SpriteComponent) as SpriteComponent[];
-      sprites.forEach((sprite: SpriteComponent) => {
-        this.objectRenderer.render(
-          sprite as TileRenderable,
-          sprite.host.point,
-          view,
-          sprite.meta
-        );
-      });
-    });
-    this.enemies.forEach((component: CombatEnemyComponent) => {
-      this.objectRenderer.render(
-        component,
-        component.renderPoint || component.point,
-        view,
-        component.meta
-      );
-      const sprites: SpriteComponent[] = component.findBehaviors(SpriteComponent);
-      sprites.forEach((sprite: SpriteComponent) => {
-        this.objectRenderer.render(
-          sprite as TileRenderable,
-          sprite.host.point,
-          view,
-          sprite.meta
-        );
-      });
-    });
-  }
-
-  afterFrame(view: SceneView, elapsed: number) {
-    // Nope
-  }
-
-  /**
-   * Custom track by function for combatants. We need to keep them around by their
-   * entity instance id so that the components are not created/destroyed as their
-   * underlying model changes. This is important because these components instances
-   * are passed through the combat state machine and their references must remain valid.
-   */
-  combatTrackBy(index: number, item: CombatantTypes): any {
-    return item.eid;
-  }
-}
