+++ src/app/routes/combat/combat-player.component.html
@@ -1,0 +1,7 @@
+<animated-behavior #animation></animated-behavior>
+<combat-attack-behavior [combat]="combat" #attack></combat-attack-behavior>
+<combat-magic-behavior [combat]="combat" #magic></combat-magic-behavior>
+<combat-item-behavior [combat]="combat" #item></combat-item-behavior>
+<combat-run-behavior [combat]="combat" #run></combat-run-behavior>
+<combat-guard-behavior [combat]="combat" #guard></combat-guard-behavior>
+<ng-content></ng-content>
