+++ src/app/routes/combat/states/combat.machine.ts
@@ -43,8 +43,8 @@ import { GameEntityObject } from '../../../scene/objects/game-entity-object';
 import { Scene } from '../../../scene/scene';
 import { SceneView } from '../../../scene/scene-view';
 import { GameWorld } from '../../../services/game-world';
-import { CombatEnemyComponent } from '../combat-enemy.entity';
-import { CombatPlayerComponent } from '../combat-player.entity';
+import { CombatEnemyComponent } from '../combat-enemy.component';
+import { CombatPlayerComponent } from '../combat-player.component';
 import { CombatAttackSummary, CombatSceneClick, IPlayerAction } from '../combat.types';
 import { CombatMachineState } from './combat-base.state';
 import { CombatDefeatSummary } from './combat-defeat.state';
